# src/swagger.yaml
...
openapi: 3.0.0
info:
  title: Express Template
  version: 1.0.0
  description: API documentation for a basic Express Template
servers:
  - url: http://localhost:4000
    description: Local server
paths:
...
  /auth/login:
    post:
      summary: "Login"
      description: "Authenticate existing user"
      operationId: "loginUser"
      tags: [Auth]
      requestBody:
        description: "User login credentials"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        "200":
          description: "Login successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      username:
                        type: string
                      id:
                        type: string
                  token:
                    type: string